[api]: main and stop functions moved into separate file
